+++ AlphaFS/Filesystem/File Class/File.GetFileSystemEntryInfo.cs
@@ -35,7 +35,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static FileSystemEntryInfo GetFileSystemEntryInfo(string path, PathFormat pathFormat)
       {
-         return GetFileSystemEntryInfoInternal(null, path, false, pathFormat);
+         return GetFileSystemEntryInfoInternal(false, null, path, false, pathFormat);
       }
 
       /// <summary>[AlphaFS] Gets the <see cref="FileSystemEntryInfo"/> of the file on the path.</summary>
@@ -44,7 +44,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static FileSystemEntryInfo GetFileSystemEntryInfo(string path)
       {
-         return GetFileSystemEntryInfoInternal(null, path, false, PathFormat.RelativePath);
+         return GetFileSystemEntryInfoInternal(false, null, path, false, PathFormat.RelativePath);
       }
 
       /// <summary>[AlphaFS] Gets the <see cref="FileSystemEntryInfo"/> of the file on the path.</summary>
@@ -55,7 +55,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static FileSystemEntryInfo GetFileSystemEntryInfo(KernelTransaction transaction, string path, PathFormat pathFormat)
       {
-         return GetFileSystemEntryInfoInternal(transaction, path, false, pathFormat);
+         return GetFileSystemEntryInfoInternal(false, transaction, path, false, pathFormat);
       }
 
       /// <summary>[AlphaFS] Gets the <see cref="FileSystemEntryInfo"/> of the file on the path.</summary>
@@ -65,7 +65,7 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public static FileSystemEntryInfo GetFileSystemEntryInfo(KernelTransaction transaction, string path)
       {
-         return GetFileSystemEntryInfoInternal(transaction, path, false, PathFormat.RelativePath);
+         return GetFileSystemEntryInfoInternal(false, transaction, path, false, PathFormat.RelativePath);
       }
 
       #endregion // GetFileSystemEntry
@@ -77,6 +77,7 @@ namespace Alphaleonis.Win32.Filesystem
       /// <remarks>BasicSearch <see cref="NativeMethods.FindExInfoLevels.Basic"/> and LargeCache <see cref="NativeMethods.FindExAdditionalFlags.LargeFetch"/> are used by default, if possible.</remarks>
       /// <exception cref="ArgumentException">The path parameter contains invalid characters, is empty, or contains only white spaces.</exception>
       /// <exception cref="ArgumentNullException">path is <c>null</c>.</exception>
+      /// <param name="isFolder">Specifies that <paramref name="path"/> is a file or directory.</param>
       /// <param name="transaction">The transaction.</param>
       /// <param name="path">The path to the file or directory.</param>
       /// <param name="continueOnException">
@@ -85,12 +86,12 @@ namespace Alphaleonis.Win32.Filesystem
       /// </param>
       /// <param name="pathFormat">Indicates the format of the path parameter(s).</param>
       [SecurityCritical]
-      internal static FileSystemEntryInfo GetFileSystemEntryInfoInternal(KernelTransaction transaction, string path, bool continueOnException, PathFormat pathFormat)
+      internal static FileSystemEntryInfo GetFileSystemEntryInfoInternal(bool isFolder, KernelTransaction transaction, string path, bool continueOnException, PathFormat pathFormat)
       {
          // Enable BasicSearch and LargeCache by default.
          var options = DirectoryEnumerationOptions.BasicSearch | DirectoryEnumerationOptions.LargeCache | (continueOnException ? DirectoryEnumerationOptions.ContinueOnException : 0);
 
-         return (new FindFileSystemEntryInfo(false, transaction, path, Path.WildcardQuestion, options, typeof (FileSystemEntryInfo), pathFormat)).Get<FileSystemEntryInfo>();
+         return (new FindFileSystemEntryInfo(isFolder, transaction, path, Path.WildcardQuestion, options, typeof(FileSystemEntryInfo), pathFormat)).Get<FileSystemEntryInfo>();
       }
 
       #endregion // GetFileSystemEntryInfoInternal